sns_backend/requirements.txt
appdirs==1.4.3
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
selenium==3.4.1
six==1.10.0
